requirements.txt
dwave-ocean-sdk >= 6.8.0